requirements.txt

fastapi
uvicorn
pydantic
requests
pyyaml
jsonschema
